fix analoginputshv errors

Closes #153

See merge request openindus/oi-soft/oi-firmware!163

+++ components/api/Middleware/Analog/InputsHV/AnalogInputsHVControl.cpp
@@ -16,11 +16,11 @@
 
 #if !defined(MODULE_STANDALONE) && !defined(MODULE_SLAVE)
 
-int AnalogInputsHVControl::analogRead(AnalogInput_Num_t num)
+float AnalogInputsHVControl::analogRead(AnalogInput_Num_t num)
 {
     std::vector<uint8_t> msgBytes = {CONTROL_ANALOG_READ, (uint8_t)num};
     _control->ctrlRequest(msgBytes);
-    int* ret = reinterpret_cast<int*>(&msgBytes[2]);
+    float* ret = reinterpret_cast<float*>(&msgBytes[2]);
     return *ret;
 }
